Fix for possibly broken nginx-common instalation
I've moved `Dpkg::Options::="--force-confdef"` down to `nginx` installation instead of `nginx-common` only. 
My case was installing `nginx` with version specified on the box with an earlier `nginx` version already installed. 

Before that change, the recipe tried to install package `nginx-common` without specifying its version, and it was reported as up to date. 
Then, `nginx` with specific version was installed and `nginx-common` was reported as needing update. During the update the prompt asking what to do with config appeared, [like this](http://opsrobot.com/post/22784291838/chef-install-apt-packages-when-apt-requires), because the `--force-confdef` option was not passed down here, and all the installation failed at unresolved dependencies.